+++ packages/pyroscope-flamegraph/src/FlameGraph/FlameGraphRenderer.tsx
@@ -8,12 +8,13 @@
 import React, { Dispatch, SetStateAction } from 'react';
 import clsx from 'clsx';
 import { Maybe } from 'true-myth';
-import { Flamebearer, Profile } from '@pyroscope/models';
+import { createFF, Flamebearer, Profile } from '@pyroscope/models';
 import Graph from './FlameGraphComponent';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore: let's move this to typescript some time in the future
 import ProfilerTable from '../ProfilerTable';
 import Toolbar from '../Toolbar';
+import NoProfilingData from '../NoProfilingData';
 import { DefaultPalette } from './FlameGraphComponent/colorPalette';
 import styles from './FlamegraphRenderer.module.scss';
 import PyroscopeLogo from '../logo-v3-small.svg';
@@ -114,10 +115,14 @@ class FlameGraphRenderer extends React.Component<
   FlamegraphRendererProps,
   FlamegraphRendererState
 > {
-  initialFlamegraphState = {
+  resetFlamegraphState = {
     focusedNode: Maybe.nothing<Node>(),
     zoom: Maybe.nothing<Node>(),
   };
+
+  // TODO: At some point the initial state may be set via the user
+  // Eg when sharing a specific node
+  initialFlamegraphState = this.resetFlamegraphState;
 
   constructor(props: FlamegraphRendererProps) {
     super(props);
@@ -150,15 +155,25 @@ class FlameGraphRenderer extends React.Component<
     prevProps: FlamegraphRendererProps,
     prevState: FlamegraphRendererState
   ) {
-    if (prevProps.profile !== this.props.profile) {
-      this.updateFlamebearerData();
-      return;
-    }
+    // TODO: this is a slow operation
+    const prevFlame = mountFlamebearer(prevProps);
+    const currFlame = mountFlamebearer(this.props);
 
-    const previousFlamebearer = prevProps.flamebearer;
-    const actualFlamebearer = this.props.flamebearer;
-    if (previousFlamebearer !== actualFlamebearer) {
-      this.updateFlamebearerData();
+    // This is a poor heuristic, but it should work most of the times
+    if (prevFlame.numTicks !== currFlame.numTicks) {
+      const newConfigs = this.calcNewConfigs(prevFlame, currFlame);
+
+      // Batch these updates to not do unnecessary work
+      // eslint-disable-next-line react/no-did-update-set-state
+      this.setState({
+        flamebearer: currFlame,
+
+        flamegraphConfigs: {
+          ...this.state.flamegraphConfigs,
+          ...newConfigs,
+        },
+      });
+      return;
     }
 
     // flamegraph configs changed
@@ -166,6 +181,46 @@ class FlameGraphRenderer extends React.Component<
       this.updateFlamegraphDirtiness();
     }
   }
+
+  // Calculate what should be the new configs
+  // It checks if the zoom/selectNode still points to the same node
+  // If not, it resets to the resetFlamegraphState
+  calcNewConfigs = (prevFlame: Flamebearer, currFlame: Flamebearer) => {
+    const newConfigs = this.state.flamegraphConfigs;
+
+    // This is a simple heuristic based on the name
+    // It does not account for eg recursive calls
+    const isSameNode = (f: Flamebearer, f2: Flamebearer, s: Maybe<Node>) => {
+      // TODO: don't use createFF directly
+      const getBarName = (f: Flamebearer, i: number, j: number) => {
+        return f.names[createFF(f.format).getBarName(f.levels[i], j)];
+      };
+
+      // No node is technically the same node
+      if (s.isNothing) {
+        return true;
+      }
+
+      return (
+        getBarName(f, s.value.i, s.value.j) ===
+        getBarName(f2, s.value.i, s.value.j)
+      );
+    };
+
+    // Reset zoom
+    const currZoom = this.state.flamegraphConfigs.zoom;
+    if (!isSameNode(prevFlame, currFlame, currZoom)) {
+      newConfigs.zoom = this.resetFlamegraphState.zoom;
+    }
+
+    // Reset focused node
+    const currFocusedNode = this.state.flamegraphConfigs.focusedNode;
+    if (!isSameNode(prevFlame, currFlame, currFocusedNode)) {
+      newConfigs.focusedNode = this.resetFlamegraphState.focusedNode;
+    }
+
+    return newConfigs;
+  };
 
   handleSearchChange = (e: string) => {
     this.setState({
@@ -293,11 +348,6 @@ class FlameGraphRenderer extends React.Component<
     return this.props.showToolbar !== undefined ? this.props.showToolbar : true;
   }
 
-  updateFlamebearerData() {
-    const flamebearer = mountFlamebearer(this.props);
-    this.setState({ flamebearer });
-  }
-
   render = () => {
     // This is necessary because the order switches depending on single vs comparison view
     const tablePane = (
@@ -305,11 +355,6 @@ class FlameGraphRenderer extends React.Component<
         key="table-pane"
         className={clsx(
           styles.tablePane,
-          {
-            [styles.hidden]:
-              !this.state.flamebearer ||
-              this.state.flamebearer.names.length <= 1,
-          },
           this.state.panesOrientation === 'vertical'
             ? styles.vertical
             : styles.horizontal
@@ -333,9 +378,6 @@ class FlameGraphRenderer extends React.Component<
         />
       </div>
     );
-    const dataExists =
-      this.state.view !== 'table' ||
-      (this.state.flamebearer && this.state.flamebearer.names.length <= 1);
 
     //    const flamegraphDataTestId = figureFlamegraphDataTestId(
     //      this.props.viewType,
@@ -344,31 +386,32 @@ class FlameGraphRenderer extends React.Component<
 
     const toolbarVisible = this.shouldShowToolbar();
 
-    const flameGraphPane =
-      this.state.flamebearer && dataExists ? (
-        <Graph
-          key="flamegraph-pane"
-          // data-testid={flamegraphDataTestId}
-          flamebearer={this.state.flamebearer}
-          ExportData={this.props.ExportData || <></>}
-          highlightQuery={this.state.highlightQuery}
-          fitMode={this.state.fitMode}
-          zoom={this.state.flamegraphConfigs.zoom}
-          focusedNode={this.state.flamegraphConfigs.focusedNode}
-          onZoom={this.onFlamegraphZoom}
-          onFocusOnNode={this.onFocusOnNode}
-          onReset={this.onReset}
-          isDirty={this.isDirty}
-          palette={this.state.palette}
-          toolbarVisible={toolbarVisible}
-          setPalette={(p) =>
-            this.setState({
-              palette: p,
-            })
-          }
-        />
-      ) : null;
+    const flameGraphPane = (
+      <Graph
+        key="flamegraph-pane"
+        // data-testid={flamegraphDataTestId}
+        flamebearer={this.state.flamebearer}
+        ExportData={this.props.ExportData || <></>}
+        highlightQuery={this.state.highlightQuery}
+        fitMode={this.state.fitMode}
+        zoom={this.state.flamegraphConfigs.zoom}
+        focusedNode={this.state.flamegraphConfigs.focusedNode}
+        onZoom={this.onFlamegraphZoom}
+        onFocusOnNode={this.onFocusOnNode}
+        onReset={this.onReset}
+        isDirty={this.isDirty}
+        palette={this.state.palette}
+        toolbarVisible={toolbarVisible}
+        setPalette={(p) =>
+          this.setState({
+            palette: p,
+          })
+        }
+      />
+    );
 
+    const dataUnavailable =
+      !this.state.flamebearer || this.state.flamebearer.names.length <= 1;
     const panes = decidePanesOrder(this.state.view, flameGraphPane, tablePane);
 
     return (
@@ -394,9 +437,7 @@ class FlameGraphRenderer extends React.Component<
               isFlamegraphDirty={this.state.isFlamegraphDirty}
               selectedNode={this.state.flamegraphConfigs.zoom}
               highlightQuery={this.state.highlightQuery}
-              onFocusOnSubtree={(i, j) => {
-                this.onFocusOnNode(i, j);
-              }}
+              onFocusOnSubtree={this.onFocusOnNode}
             />
           )}
           {this.props.children}
@@ -409,7 +450,7 @@ class FlameGraphRenderer extends React.Component<
               styles.panesWrapper
             )}`}
           >
-            {panes.map((pane) => pane)}
+            {dataUnavailable ? <NoProfilingData /> : panes.map((pane) => pane)}
           </div>
         </div>